Add 'manual gg' feature
Instead of having to reach over to press T, then type "gg", then press Enter - a whole strenuous three step process of navigating across the keyboard like some sort of gymnist and having to take your hand off the mouse - all you have to do is press the G key twice and "gg" will be sent in chat.

This is a feature for those that don't like the disingenuous nature of "auto gg" features but still want to show good sportsmanship.