Add an emote selector for Hypixel emotes
The code has been designed to allow for user-defined emotes, but users have no way to actually define such emotes. Maybe this can be done later, but at least for the time being, this isn't necessary at all.

The emote selector can be accessed, by default, with Ctrl + E with the chat open, navigated with the arrow keys, and using Enter to insert the emote at the cursor position.

Some desirables, but not important:
- Favourite emote feature, always at the top or in its own dedicated section in the widget
- Dynamic positioning of the widget if opened away from the chat window (e.g. move it right if opened while typing closer to the right)
- Custom emote definitions
- A little bit more configurability for the widget