Remove configurable keybinds from chat features
Minecraft key mappings have a limitation of only allowing one key mapping to function for each unique key. This means that if a key is configured to Tab (the default for switching channels), you can no longer open the player list unless that key is changed. Additionally, vanilla keybinds do not support control codes/modifiers, which imposes an additional restriction of all keys requiring Ctrl in the mod, which is a bit annoying.

There are mods such as AMECS that alleviate these limitations, but AMECS is not updated for 1.21. Forge also has this built-in natively, but Fabric's core principle is to not modify the client at all, just provide a modding layer and some APIs.

I don't want to reimplement AMECS' functionality into MEH, so until there's a better alternative or solution, I'm just going to have to hard-code these types of keys. I'll still add keybindings for non-chat-related features though because that just makes sense imo.