Add a new feature to retexture Hypixel lobby fishing rods
The textures are not owned by me, they were created by the community under the name "Fishers' Delight", which is a resource pack for Minecraft 1.8 with MCPatcher. All credits for the textures go towards the creator of Fishers' Delight. I do not claim any credit. A disclaimer has been added to the README and I will remove these textures if asked.

https://hypixel.net/threads/fishers-delight-a-main-lobby-fishing-pack.5411511/