Simplify mnemonic registration a bit. Generate language entries
Unfortunately registries are a bit complex and probably way too overkill for this anyways. When using registries, the data generator is run too late and the registries by this point are frozen, so we can't create new registries (and consequently, the data generator fails). The solution for this is to use a BootstrapContext object in the DataGeneratorEntrypoint, but that also means I need a BootstrapContext object when registering entries on the client... which I don't know how to do.

Overall, to save myself the headache, let's just avoid registries for now. As much as I'd love to use registries for this, emotes, chat channels, and more, I think the better option is to just not for now.