Abandon data generators in favour of manual JSON files
Unfortunately due to the way data generators are executed, they need access to server-sided resources and thus forces the mod to be unable to mark itself as client-sided only. The data generator entrypoint can only be run on the logical server, so if the mod is marked as client only, the data generator is not found and fails to run.

It seems a lot of popular client-sided mods also don't use data generators (likely for the same reason). MEH doesn't generate enough data for a data generator to really be worth the effort. So let's just do this manually instead.