Support for Amecs keybinds
Amecs has not yet updated to 1.21, but there is a fork called "Too Many Shortcuts". I don't personally intend on using these mods because the default keybinds are keybinds that I would use anyways... but I do still want to have configurable keybinds for those that may have conflicts.

The reason Amecs is a requirement for these keybinds specifically is because they require the Ctrl keycode in order for them to be used without typing characters in the chat screen. I'd like to eventually have it so that any keybind doesn't write the character in the chat screen, but for now that's not a massive enough issue.

Despite Amecs (or any derivative) being a requirement for these keybinds, it's still an optional dependency. Additionally, this doesn't mean that MEH will never have configurable keybinds. It's just that all the keybinds that exist so far are done in the chat screen and require Ctrl, which vanilla does not support.