Introduce "focus mode" for chat channels
This commit adds a new mode for channels called "focus mode" which can be toggled, by default, with Ctrl + F while the chat screen is open. This can be changed if Amecs is installed. While in focus mode, only messages related to that chat channel will be shown in the chat.

The following changes have been made to accommodate and accompany these new changes:
- Added a new (root) configuration option to change how many messages are remembered (but not drawn) by the client, which defaults to 250 (from vanilla's 100)
- Added configuration options under known channels to change the filters of default channels (i.e., by default, the party and guild channels)
- Rendered an icon next to a channel if it is in focus mode, as well as a tooltip when hovered explaining how to disable it
- Abstracted out chat message parsing into a new ChatExtractor system which allows for matching text and extracting specific pieces of the message if it matches
- Fixed ChannelSelector#getPreviousChannel() returning the incorrect channel