Add a feature to retexture Murder Mystery's murderer knives
I think that the murderer knives in Murder Mystery leave a lot to be desired. I would like to add custom textures to all cosmetic knife skins in this game mode.

This will probably stay off of the main branch for a little bit but I want to tackle this as a bit of a side project to better my pixel art abilities.