Improve how server ids and locations are obtained, add event
MEH now optionally depends on the official Hypixel Mod API (https://modrinth.com/mod/hypixel-mod-api) and will provide some more accurate information about the server and environment the player is connected to. Specifically, with the mod installed, MEH can now determine whether or not the client is connected to the production, beta, or test environment, and more accurately determine the server type and whether or not the player is in a lobby.

The server type and lobby state can be determined via the scoreboard, and MEH will make a best-effort attempt to do so if the Hypixel Mod API isn't installed, but the HPMAPI makes these more accurate and also provides the environment, which you can't obtain from the scoreboard reliably.

HypixelServerEvents.SERVER_LOCATION_CHANGE adds a hook for when the client joins a new server and can compare it to the server they came from.