Add new feature to auto enable night vision in houses
This feature will only work if the housing uses a new Housing feature to send command suggestions to the client. It's done this way so that the client doesn't just spam a bunch of commands when joining, throwing shit at the wall to see what sticks.

The existing "auto disable housing flight" feature has been moved into a category alongside the auto night vision toggle under a "housing" category.